Task-06/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
marshmallow==3.20.2
orjson
redis